--- python/rubbish/__init__.py
@@ -1,3 +1,0 @@
-from . import common
-from . import admin
-from . import client
