--- python/rubbish/common/__init__.py
@@ -1,3 +1,0 @@
-from .orm import *
-from .consts import *
-from .db_ops import *
