+++ python/rubbish_geo_admin/setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup, find_packages
 
 setup(
-    name='rubbish',
+    name='rubbish_geo_admin',
     version='0.1',
     packages=find_packages(),
     install_requires=[
