+++ python/rubbish_geo_client/ops.py
@@ -12,9 +12,9 @@ import geoalchemy2
 from geoalchemy2.shape import to_shape
 from scipy.stats import shapiro
 
-from rubbish.common.db_ops import db_sessionmaker
-from rubbish.common.orm import Pickup, Centerline, BlockfaceStatistic, Sector
-from rubbish.common.consts import RUBBISH_TYPES, RUBBISH_TYPE_MAP
+from rubbish_geo_common.db_ops import db_sessionmaker
+from rubbish_geo_common.orm import Pickup, Centerline, BlockfaceStatistic, Sector
+from rubbish_geo_common.consts import RUBBISH_TYPES, RUBBISH_TYPE_MAP
 
 def point_side_of_centerline(point_geom, centerline_geom):
     """
