+++ python/rubbish_geo_common/consts.py
